+++ test/test.sh
@@ -68,13 +68,13 @@ ${ROOT}/bin/flatcc -I ${ROOT}/test/monster_test -a \
 cd ${TMP}/monster_test_main
 cp ${ROOT}/test/monster_test/monster_test.c .
 $CC -g -I ${ROOT}/include monster_test.c \
-    ${ROOT}/lib/libflatccrt.a -o monster_test
+    ${ROOT}/lib/libflatccrt_d.a -o monster_test_d
 $CC -O3 -DNDEBUG -DFLATBUFFERS_BENCHMARK -I ${ROOT}/include monster_test.c \
-    ${ROOT}/lib/libflatccrt.a -o monster_test_optimized
-./monster_test
+    ${ROOT}/lib/libflatccrt.a -o monster_test
+./monster_test_d
 
 echo "running optimized version of main monster test"
-./monster_test_optimized
+./monster_test
 
 echo "running reflection test"
 ${ROOT}/test/reflection_test//reflection_test.sh